Add FormModalUtils class instead of inheritance
