+++ package.json
@@ -94,13 +94,13 @@
     "@babel/preset-typescript": "^7.3.3",
     "@mighty-justice/tester": "^1.3.0",
     "@mighty-justice/tslint-config": "^0.3.0",
-    "@storybook/addon-actions": "^5.0.6",
-    "@storybook/addon-info": "^5.0.6",
-    "@storybook/addon-links": "^5.0.6",
-    "@storybook/addons": "^5.0.6",
-    "@storybook/react": "^5.0.6",
+    "@storybook/addon-actions": "^5.0.11",
+    "@storybook/addon-info": "^5.0.11",
+    "@storybook/addon-links": "^5.0.11",
+    "@storybook/addons": "^5.0.11",
+    "@storybook/react": "^5.0.11",
     "@storybook/storybook-deployer": "^2.8.1",
-    "@storybook/theming": "^5.0.6",
+    "@storybook/theming": "^5.0.11",
     "@types/classnames": "^2.2.7",
     "@types/enzyme": "^3.9.1",
     "@types/enzyme-adapter-react-16": "^1.0.5",
@@ -110,7 +110,7 @@
     "@types/lodash": "^4.14.123",
     "@types/react": "16.8.8",
     "@types/rosie": "^0.0.36",
-    "@types/storybook-readme": "^5.0.1",
+    "@types/storybook-readme": "^5.0.3",
     "@types/storybook__addon-actions": "^3.4.2",
     "@types/storybook__addon-info": "^4.1.1",
     "@types/storybook__react": "^4.0.1",
