+++ src/components/FormDrawer.tsx
@@ -1,21 +1,31 @@
-import React from 'react';
+import React, { Component } from 'react';
 import { observer } from 'mobx-react';
 import autoBindMethods from 'class-autobind-decorator';
 import cx from 'classnames';
 
 import * as Antd from 'antd';
 
+import { ISharedFormModalProps } from '../props';
+import FormModalUtils from '../utilities/FormModalUtils';
+
 import Form from './Form';
-import FormModal from './FormModal';
 
 @autoBindMethods
 @observer
-class FormDrawer extends FormModal {
+class FormDrawer extends Component<ISharedFormModalProps> {
+  private formModalUtils: FormModalUtils;
+
+  public constructor (props: ISharedFormModalProps) {
+    super(props);
+    this.formModalUtils = new FormModalUtils(props);
+  }
+
   public render () {
     const { className, title, width } = this.props
+      , { isVisible, onCancel, onSuccess, formProps } = this.formModalUtils
       , drawerClassName = cx('mfa-form-drawer', className || null);
 
-    if (!this.isVisible) { return null; }
+    if (!isVisible) { return null; }
 
     return (
       <Antd.Drawer
@@ -23,7 +33,7 @@ class FormDrawer extends FormModal {
         closable
         destroyOnClose
         maskClosable={false}
-        onClose={this.onCancel}
+        onClose={onCancel}
         placement='right'
         title={title}
         visible={true}
@@ -32,9 +42,9 @@ class FormDrawer extends FormModal {
         {this.props.childrenBefore}
 
         <Form
-          {...this.formProps}
-          onCancel={this.onCancel}
-          onSuccess={this.onSuccess}
+          {...formProps}
+          onCancel={onCancel}
+          onSuccess={onSuccess}
         />
       </Antd.Drawer>
     );
